refactor(expert-chat): satisfy multi-session CI gates
